# Optimize GIFT-128 Bit Permutation