**Seeing ~116% performance improvement of GIFT-128 permutation, on `Intel(R) Core(TM) i5-8279U CPU @ 2.40GHz`**

> 40 rounds of GIFT-128 permutation [ **before** ]

```fish
2022-07-13T15:54:20+04:00
Running ./bench/a.out
Run on (8 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB
  L1 Instruction 32 KiB
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB
Load Average: 0.93, 1.10, 1.44
----------------------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------
bench_gift_cofb::gift_permute       2943 ns         2941 ns       239834 bytes_per_second=5.18864M/s
```

> 40 rounds of GIFT-128 permutation [ **after** ]

```fish
2022-07-13T15:54:59+04:00
Running ./bench/a.out
Run on (8 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB
  L1 Instruction 32 KiB
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB
Load Average: 0.78, 1.04, 1.41
----------------------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------
bench_gift_cofb::gift_permute       1362 ns         1361 ns       508388 bytes_per_second=11.2079M/s
```